# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Callora API",
    "version": "1.0.0",
    "description": "API contract covering Callora backend billing, usage, and developer routes."
  },
  "paths": {
...
    "/api/billing/deduct": {
      "post": {
        "summary": "Deduct billing balance for an API call",
        "description": "Deducts USDC from the user's vault balance to pay for an API call. Authenticated and idempotent.",
        "externalDocs": {
          "description": "SDK idempotency contract and retry guidance",
          "url": "https://callora.com/sdk/billing-deduct.md"
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingDeductRequest"
              },
              "examples": {
                "deductRequest": {
                  "summary": "Deduct billing request",
                  "value": {
                    "requestId": "req-123e4567-e89b-12d3-a456-426614174000",
                    "developerId": "dev-123",
                    "apiId": "api-123",
                    "endpointId": "endpoint-456",
                    "apiKeyId": "key-789",
                    "amountUsdc": "0.01",
                    "idempotencyKey": "idem-abc123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Billing deduction successfully processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingDeductResponse"
                },
                "examples": {
                  "success": {
                    "summary": "Successful deduction",
                    "value": {
                      "success": true,
                      "usageEventId": "evt-123e4567-e89b-12d3-a456-426614174000",
                      "stellarTxHash": "abc123def456...",
                      "alreadyProcessed": false
                    }
                  },
                  "alreadyProcessed": {
                    "summary": "Already processed (idempotent)",
                    "value": {
                      "success": true,
                      "usageEventId": "evt-123e4567-e89b-12d3-a456-426614174000",
                      "stellarTxHash": "abc123def456...",
                      "alreadyProcessed": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidAmount": {
                    "summary": "Invalid amount format",
                    "value": {
                      "code": "BAD_REQUEST",
                      "message": "amountUsdc must be a positive number with at most 7 decimal places",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Missing or invalid authentication",
                    "value": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication required",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment required (insufficient balance)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "insufficientBalance": {
                    "summary": "Insufficient balance",
                    "value": {
                      "code": "INSUFFICIENT_BALANCE",
                      "message": "Vault balance too low for deduction",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "idempotencyConflict": {
                    "summary": "Idempotency key already used with different parameters",
                    "value": {
                      "code": "IDEMPOTENCY_CONFLICT",
                      "message": "Idempotency key conflict: different request parameters",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/billing/deduct/bulk": {
      "post": {
        "summary": "Deduct billing balance for up to 100 API calls in one batch",
        "description": "Creates usage events for up to 100 billing entries in a single database transaction, then performs one aggregate on-chain deduction for the newly inserted entries. Existing requestIds are treated idempotently and are returned in the per-entry results without being charged again.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingBulkDeductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk billing deduction successfully processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingBulkDeductResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "description": "Seconds until the rate limit expires"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "rateLimited": {
                    "summary": "Too many requests",
                    "value": {
                      "code": "TOO_MANY_REQUESTS",
                      "message": "Too Many Requests",
                      "requestId": "req-abc123"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing/disputes": {
      "post": {
        "summary": "Open a dispute",
        "description": "Developer opens a dispute against a billing charge (usage event). One open dispute is allowed per usage_event_id.",
        "tags": [
          "Billing",
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDisputeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dispute opened successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Dispute already exists for this usage_event_id"
          }
        }
      },
      "get": {
        "summary": "List own disputes",
        "description": "Returns all disputes opened by the authenticated developer.",
        "tags": [
          "Billing",
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of disputes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "disputes",
                    "total"
                  ],
                  "properties": {
                    "disputes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Dispute"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-api-key"
      }
    },
    "schemas": {
...
      "BillingDeductRequest": {
        "type": "object",
        "required": [
          "requestId",
          "apiId",
          "endpointId",
          "apiKeyId",
          "amountUsdc"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "developerId": {
            "type": "string",
            "description": "Optional developer identifier. If provided, it must be a non-empty string. When omitted, the authenticated user ID is used."
          },
          "apiId": {
            "type": "string"
          },
          "endpointId": {
            "type": "string"
          },
          "apiKeyId": {
            "type": "string"
          },
          "amountUsdc": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          }
        }
      },
      "BillingDeductResponse": {
        "type": "object",
        "required": [
          "success",
          "usageEventId",
          "stellarTxHash",
          "alreadyProcessed"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "usageEventId": {
            "type": "string"
          },
          "stellarTxHash": {
            "type": "string"
          },
          "alreadyProcessed": {
            "type": "boolean"
          }
        }
      },
      "BillingBulkDeductEntry": {
        "type": "object",
        "required": [
          "events",
          "stats",
          "period"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "apiId": {
            "type": "string"
          },
          "endpointId": {
            "type": "string"
          },
          "apiKeyId": {
            "type": "string"
          },
          "amountUsdc": {
            "type": "string",
            "description": "Positive decimal string with up to 7 fractional digits"
          }
        }
      },
      "BillingBulkDeductRequest": {
        "type": "object",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/BillingBulkDeductEntry"
            }
          },
          "idempotencyKey": {
            "type": "string"
          }
        }
      },
      "BillingBulkDeductEntryResult": {
        "type": "object",
        "required": [
          "requestId",
          "usageEventId",
          "alreadyProcessed",
          "deductionApplied",
          "reconciliationRequired"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "usageEventId": {
            "type": "string"
          },
          "stellarTxHash": {
            "type": "string"
          },
          "alreadyProcessed": {
            "type": "boolean"
          },
          "deductionApplied": {
            "type": "boolean"
          },
          "reconciliationRequired": {
            "type": "boolean"
          }
        }
      },
      "BillingBulkDeductResponse": {
        "type": "object",
        "required": [
          "success",
          "entryCount",
          "deductedCount",
          "totalDeductedAmountUsdc",
          "results"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "entryCount": {
            "type": "integer"
          },
          "deductedCount": {
            "type": "integer"
          },
          "totalDeductedAmountUsdc": {
            "type": "string"
          },
          "stellarTxHash": {
            "type": "string"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillingBulkDeductEntryResult"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "PAYMENT_REQUIRED",
          "TOO_MANY_REQUESTS",
          "CONFLICT",
          "INTERNAL_SERVER_ERROR",
          "BAD_GATEWAY",
          "SERVICE_UNAVAILABLE",
          "GATEWAY_TIMEOUT",
          "VALIDATION_ERROR",
          "INVALID_BODY",
          "INVALID_QUERY",
          "INVALID_PARAMS",
          "INVALID_VALUE",
          "GATEWAY_AUTH_CONTEXT_MISSING",
          "UPSTREAM_TARGET_BLOCKED",
          "INSUFFICIENT_BALANCE",
          "SOROBAN_RPC_TIMEOUT",
          "SOROBAN_RPC_ERROR",
          "BILLING_DEDUCTION_FAILED",
          "BILLING_REQUEST_NOT_FOUND",
          "DEVELOPER_NOT_FOUND",
          "API_ACCESS_FORBIDDEN",
          "API_KEY_NOT_FOUND",
          "API_KEY_FORBIDDEN",
          "MISSING_REFRESH_TOKEN",
          "INVALID_REFRESH_TOKEN",
          "REVOKED_TOKEN",
          "EXPIRED_TOKEN",
          "REFRESH_FAILED",
          "REVOKE_FAILED",
          "NOT_AUTHENTICATED",
          "TOKEN_INFO_FAILED",
          "VAULT_NOT_FOUND",
          "VAULT_BALANCE_RETRIEVAL_FAILED",
          "MISSING_AMOUNT",
          "INVALID_AMOUNT_TYPE",
          "INVALID_AMOUNT_FORMAT",
          "INVALID_NETWORK",
          "NETWORK_MISMATCH",
          "INVALID_SOURCE_ACCOUNT",
          "INVALID_TRANSACTION_INPUT",
          "SOURCE_ACCOUNT_NOT_FOUND",
          "INVALID_CONTRACT_ID",
          "NETWORK_UNAVAILABLE",
          "TRANSACTION_BUILD_FAILED",
          "INTERNAL_ERROR",
          "INVALID_WEBHOOK_REGISTRATION",
          "INVALID_WEBHOOK_EVENT_TYPES",
          "WEBHOOK_NOT_FOUND",
          "INVALID_WEBHOOK_URL",
          "WEBHOOK_URL_VALIDATION_FAILED",
          "MISSING_WEBHOOK_SIGNATURE_HEADERS",
          "INVALID_WEBHOOK_TIMESTAMP",
          "WEBHOOK_TIMESTAMP_OUT_OF_WINDOW",
          "MALFORMED_WEBHOOK_SIGNATURE",
          "INVALID_WEBHOOK_SIGNATURE",
          "INVALID_DELIVERY_ID",
          "INVALID_RETRY_POLICY",
          "DLQ_ENTRY_NOT_FOUND",
          "INVALID_IP_FORMAT",
          "IP_NOT_ALLOWED",
          "DATABASE_NOT_AVAILABLE",
          "IDEMPOTENCY_CONFLICT",
          "IDEMPOTENCY_IN_PROGRESS",
          "SIMULATION_FAILED",
          "INVALID_AUTH_HEADER",
          "MISSING_TOKEN",
          "INVALID_TOKEN",
          "MISSING_CLAIMS",
          "TOKEN_EXPIRED",
          "TOKEN_NOT_ACTIVE",
          "QUOTA_REQUEST_NOT_FOUND",
          "QUOTA_REQUEST_ALREADY_RESOLVED",
          "INVALID_QUOTA_REQUEST",
          "REQUEST_TIMEOUT",
          "REQUEST_BODY_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "UNPROCESSABLE_ENTITY",
          "USAGE_AGGREGATE_NOT_FOUND",
          "INVALID_EXPORT_SCHEDULE",
          "EXPORT_SCHEDULE_NOT_FOUND",
          "MISSING_AUTH_FIELDS",
          "AUTH_NOT_IMPLEMENTED",
          "COMPONENT_NOT_CONFIGURED"
        ],
        "description": "Canonical Callora backend error code."
      },
      "Dispute": {
        "type": "object",
        "required": [
          "id",
          "usage_event_id",
          "opened_by",
          "reason",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "usage_event_id": {
            "type": "string"
          },
          "opened_by": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "REFUNDED",
              "UPHELD"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolved_by": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "OpenDisputeRequest": {
        "type": "object",
        "required": [
          "usage_event_id",
          "reason"
        ],
        "properties": {
          "usage_event_id": {
            "type": "string",
            "minLength": 1
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        }
      },
...
    }
  }
}